refactor(chatbot): Add seo to chatbot and make the page static
